feat: add action class as builder for a executable action
Introduce Action class acting as a builder for creating an executable
action. Define functions setInputSchema() and setActionFn() to set the
zod input schema and the action function to execute after validation
respectively. The setActionFn() function returns another function that
requires the input parameter as the inferred zod type of the input
schema.

Define a test suite for testing the action class with two unit tests to
check the Action class with valid input and another to check whether the
executable function throws if the action is undefined.