feat: add validators for action
Add types ValidationResult and ValidationFn to define the return type of
validators and their function signatures.

Update Action class to include an array of ValidationFn which are
executed before the action is executed. Each validator can add something
to the context. Each validator can support error codes and a payload
corresponding to that error code and if a validator is unsuccessful then
the action terminates and returns the error code along with the payload.